@@ -30,8 +30,12 @@ public class CategoryController {
     @GetMapping("/page")
     public ResponseEntity<List<CategoryDTO>> getAllCategoriesByPage(@RequestParam(value = "offset", defaultValue = "0") @Min(0) Integer offset,
                                                                     @RequestParam(value = "limit", defaultValue = "10") @Min(1) @Max(100) Integer limit,
-                                                                    @RequestParam(value = "sort", required = false) String sortField) {
-        Page<CategoryDTO> page = categoryService.getAllCategoriesByPage(PageRequest.of(offset, limit, Sort.by(Sort.Direction.ASC, sortField)));
+                                                                    @RequestParam(value = "sort", required = false) String sortField,
+                                                                    @RequestParam(value = "direction", defaultValue = "asc") String sortDirection){
+        Sort.Direction direction = Sort.Direction.fromString(sortDirection.toLowerCase());
+        Page<CategoryDTO> page = categoryService.getAllCategoriesByPage(
+                PageRequest.of(offset, limit, Sort.by(direction, sortField))
+        );
         List<CategoryDTO> categories = page.getContent();
         return ResponseEntity.ok(categories);
     }
