@@ -9,6 +9,6 @@ import java.util.Optional;
 
 
 public interface UserRepository  extends JpaRepository <User,Long> {
-    User findByLogin(String login);
-    User findByEmail(String login);
+    Optional<User> findByLogin(String login);
+    Optional<User> findByEmail(String login);
 }
